web/moca-grill/src/backend/requirements.txt
asgiref==3.8.1
Django==4.2.11
django-rest-registration==0.9.0
djangorestframework==3.15.2
drf-nested-routers==0.94.1
gunicorn==23.0.0
packaging==24.1
sqlparse==0.5.1
